backend/requirements-simple.txt
Flask
Flask-CORS
Pillow
numpy
trimesh
supabase
python-dotenv


